fix: validating empty regexp path params
Previous implementation on allowing empty path params was broken and it
allowed empty path params to pass through even when regexp didn't allow
that. Previous tests gave false positives due to the order of the routes
inserted, but the actual implementation was faulty and problem was
revealed by just changing the order of the routes inserted.

This fix should solve that problem and also added additional tests to
ensure more cases.

One existing test mux test was also broken since after the
implementation was fixed it started to receive the correct "404"
response.